Add AIX ppc64 build
