+++ .github/workflows/build-and-test.yml
@@ -148,7 +148,17 @@ jobs:
   cross-compile:
     strategy:
       matrix:
-        SYS_BINARIES: [ "binaries-darwin_amd64", "binaries-darwin_arm64", "binaries-linux_amd64", "binaries-linux_arm64", "binaries-windows_amd64", "binaries-windows_arm64", "binaries-linux_ppc64le", "binaries-linux_s390x", "binaries-windows_386" ]
+        SYS_BINARIES:
+          - binaries-aix_ppc64
+          - binaries-darwin_amd64
+          - binaries-darwin_arm64
+          - binaries-linux_amd64
+          - binaries-linux_arm64
+          - binaries-linux_ppc64le
+          - binaries-linux_s390x
+          - binaries-windows_386
+          - binaries-windows_amd64
+          - binaries-windows_arm64
     uses: ./.github/workflows/reusable-compile.yml
     with:
       SYS_BINARY: ${{ matrix.SYS_BINARIES }}
