+++ internal/configsource/vaultconfigsource/gcp_auth_token_aix.go
@@ -1,0 +1,29 @@
+// Copyright Splunk, Inc.
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build aix
+
+package vaultconfigsource
+
+import (
+	"fmt"
+	"runtime"
+
+	"github.com/hashicorp/vault/api"
+)
+
+func (gcp *GCPAuthentication) Token(client *api.Client) (string, error) {
+	return "", fmt.Errorf("token retrieval is not supported on %s", runtime.GOOS)
+}
