+++ internal/signalfx-agent/pkg/core/modules.go
@@ -1,3 +1,5 @@
+//go:build !aix
+
 package core
 
 // Do an import of all of the built-in observers and monitors
