+++ internal/signalfx-agent/pkg/core/modules_aix.go
@@ -1,0 +1,3 @@
+package core
+
+// No legacy monitors are supported by new platforms
